--- apps/web/src/providers/index.ts
@@ -1,3 +1,0 @@
-export { default as QueryClientProvider } from "./QueryClientProvider/QueryClientProvider";
-export { default as RainbowKitProvider } from "./RainbowKitProvider/RainbowKitProvider";
-export { default as WagmiProvider } from "./WagmiProvider/WagmiProvider";
